refactor: split bdd grid core into srp modules
