+++ crates/bitnet-bdd-grid-core/src/environment.rs
@@ -1,0 +1,40 @@
+use std::fmt;
+use std::str::FromStr;
+
+use crate::labels::normalize_label;
+
+/// Execution environment axis for BDD planning.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum ExecutionEnvironment {
+    Local,
+    Ci,
+    PreProduction,
+    Production,
+}
+
+impl fmt::Display for ExecutionEnvironment {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Local => write!(f, "local"),
+            Self::Ci => write!(f, "ci"),
+            Self::PreProduction => write!(f, "pre-prod"),
+            Self::Production => write!(f, "production"),
+        }
+    }
+}
+
+impl FromStr for ExecutionEnvironment {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match normalize_label(s).as_str() {
+            "local" | "dev" | "development" => Ok(Self::Local),
+            "ci" | "ci/cd" | "cicd" => Ok(Self::Ci),
+            "pre-prod" | "preprod" | "pre-production" | "preproduction" | "staging" => {
+                Ok(Self::PreProduction)
+            }
+            "prod" | "production" => Ok(Self::Production),
+            _ => Err("unknown execution environment"),
+        }
+    }
+}
