+++ crates/bitnet-bdd-grid-core/src/labels.rs
@@ -1,0 +1,3 @@
+pub(crate) fn normalize_label(input: &str) -> String {
+    input.trim().to_ascii_lowercase().replace(['_', ' '], "-")
+}
