+++ crates/bitnet-bdd-grid-core/src/scenario.rs
@@ -1,0 +1,53 @@
+use std::fmt;
+use std::str::FromStr;
+
+use crate::labels::normalize_label;
+
+/// Logical test scenario axis for BDD planning.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum TestingScenario {
+    Unit,
+    Integration,
+    EndToEnd,
+    Performance,
+    CrossValidation,
+    Smoke,
+    Development,
+    Debug,
+    Minimal,
+}
+
+impl fmt::Display for TestingScenario {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Unit => write!(f, "unit"),
+            Self::Integration => write!(f, "integration"),
+            Self::EndToEnd => write!(f, "e2e"),
+            Self::Performance => write!(f, "performance"),
+            Self::CrossValidation => write!(f, "crossval"),
+            Self::Smoke => write!(f, "smoke"),
+            Self::Development => write!(f, "development"),
+            Self::Debug => write!(f, "debug"),
+            Self::Minimal => write!(f, "minimal"),
+        }
+    }
+}
+
+impl FromStr for TestingScenario {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match normalize_label(s).as_str() {
+            "unit" => Ok(Self::Unit),
+            "integration" => Ok(Self::Integration),
+            "e2e" | "end-to-end" | "endtoend" => Ok(Self::EndToEnd),
+            "performance" | "perf" => Ok(Self::Performance),
+            "crossval" | "cross-validation" => Ok(Self::CrossValidation),
+            "smoke" => Ok(Self::Smoke),
+            "development" | "dev" => Ok(Self::Development),
+            "debug" => Ok(Self::Debug),
+            "minimal" | "min" => Ok(Self::Minimal),
+            _ => Err("unknown testing scenario"),
+        }
+    }
+}
